Created Phase REST editor
